fix(plugin-user-profile): CB-1399 style changes
