+++ webapp/packages/plugin-authentication/src/UserMenu/UserMenu.tsx
@@ -11,10 +11,11 @@ import styled from 'reshadow';
 
 import { topMenuStyles } from '@cloudbeaver/core-app';
 import { AuthInfoService, DATA_CONTEXT_USER } from '@cloudbeaver/core-authentication';
+import { Icon } from '@cloudbeaver/core-blocks';
 import { useService } from '@cloudbeaver/core-di';
 import { useStyles } from '@cloudbeaver/core-theming';
 import { ContextMenu } from '@cloudbeaver/core-ui';
-import { MenuBaseItem, useMenu } from '@cloudbeaver/core-view';
+import { useMenu } from '@cloudbeaver/core-view';
 
 import { UserInfo } from '../UserInfo';
 import { MENU_USER_PROFILE } from './MENU_USER_PROFILE';
@@ -25,25 +26,18 @@ export const UserMenu = observer(function UserMenu() {
   const authInfoService = useService(AuthInfoService);
   const menu = useMenu(MENU_USER_PROFILE);
 
+  menu.context.set(DATA_CONTEXT_USER, authInfoService.userInfo);
+
   if (!authInfoService.userInfo) {
     return null;
   }
 
-  const items = menu.getItems().filter(item => !item.hidden);
-
-  if (items.length === 1 && items[0] instanceof MenuBaseItem) {
-    const item = items[0];
-
-    if (item.events?.onSelect) {
-      return <UserInfo info={authInfoService.userInfo} tooltip={item.label} detached onClick={item.events.onSelect} />;
-    }
-  }
-
-  menu.context.set(DATA_CONTEXT_USER, authInfoService.userInfo);
-
   return styled(style)(
-    <ContextMenu menu={menu} style={[topMenuStyles]} rtl modal>
+    <>
       <UserInfo info={authInfoService.userInfo} />
-    </ContextMenu>
+      <ContextMenu menu={menu} style={[topMenuStyles]} rtl modal>
+        <Icon name="angle" viewBox="0 0 15 8" />
+      </ContextMenu>
+    </>
   );
 });
