+++ webapp/packages/plugin-authentication/src/UserMenu/userMenuStyles.ts
@@ -11,11 +11,17 @@ import { css } from 'reshadow';
 export const userMenuStyles = css`
   ContextMenu {
     opacity: 1;
-    padding: 0 16px !important;
+    padding: 0 !important;
     height: 48px;
     cursor: pointer;
     background: none;
     border: none;
     outline: none !important;
+    & Icon {
+        background: #47a0dd;
+        width: 16px;
+        height: 100%;
+        padding: 0 16px;
+    }
   }
 `;
