+++ webapp/packages/plugin-user-profile/src/index.ts
@@ -1,3 +1,5 @@
 import { userProfilePlugin } from './manifest';
 
 export default userProfilePlugin;
+
+export * from './UserProfileService';
